mcp-agent/requirements.txt
# MCP SDK
mcp>=1.26.0

# LLM
anthropic>=0.45.0

# Web framework
fastapi>=0.115.0
uvicorn>=0.34.0

# Data sources
pymongo>=4.10.0
redis>=5.2.0
aiokafka>=0.12.0
paho-mqtt>=2.1.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.28.0
